+++ TestWebsite/Client/Scripts/Pages/Test/Inputs/CustomOption.tsx
@@ -1,14 +1,19 @@
 ﻿import * as React from 'react'
 import { OptionProps } from 'react-select/lib/components/Option'
 import { components } from 'react-select/lib/components'
-import { SelectOption } from '@interface-technologies/iti-react'
 
-export function CustomOption(props: OptionProps<SelectOption>) {
+export interface CustomOptionType {
+    value: string
+    label: string
+    color: string
+}
+
+export function CustomOption(props: OptionProps<CustomOptionType>) {
     const { children, ...nonChildrenProps } = props
 
     return (
         <components.Option {...nonChildrenProps}>
-            <span style={{ color: 'lightseagreen' }}>~{children}~</span>
+            <span style={{ color: props.data.color }}>~{children}~</span>
         </components.Option>
     )
 }