+++ TestWebsite/Client/Scripts/Pages/Test/Inputs/SelectOptions.ts
@@ -1,4 +1,6 @@
-﻿export const colorOptions = [
+﻿import { CustomOptionType } from './CustomOption'
+
+export const colorOptions = [
     { value: 'ocean', label: 'Ocean', color: '#00B8D9' },
     { value: 'blue', label: 'Blue', color: '#0052CC', disabled: true },
     { value: 'purple', label: 'Purple', color: '#5243AA' },