+++ iti-react/Util/RoutingUtil.ts
@@ -2,7 +2,7 @@
 import { stripTrailingSlash } from 'history/PathUtils'
 import { isEqual } from 'lodash'
 
-export function locationsAreEqualIgnoringKey(a: Location, b: Location) {
+export function areLocationsEqualIgnoringKey(a: Location, b: Location) {
     return (
         a.pathname === b.pathname &&
         a.search === b.search &&